# docs/swagger.yaml
definitions:
...
  models.Discount:
    properties:
      amount:
        description: Percentage or fixed amount
        type: number
      created_at:
        type: string
      description:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_quantity:
        description: Maximum quantity for discount (0 = unlimited)
        type: integer
      min_quantity:
        description: Minimum quantity for discount
        type: integer
      name:
        type: string
      tier:
        $ref: '#/definitions/models.EventTier'
      tier_id:
        description: If discount applies to specific tier
        type: string
      type:
        description: percentage, fixed_amount, buy_x_get_y
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
      usage_limit:
        description: 0 = unlimited
        type: integer
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
  models.Event:
    properties:
      address:
        type: string
      admin_remark:
        type: string
      available:
        type: integer
      banner_image:
        type: string
      cancel_reason:
        type: string
      cancelled_at:
        type: string
      capacity:
        minimum: 1
        type: integer
      category:
        description: Array of category tags
        items:
          type: string
        type: array
      commission_rate:
        description: Platform commission percentage (0-100)
        type: number
      created_at:
        type: string
      description:
        description: HTML content
        type: string
      discounts:
        items:
          $ref: '#/definitions/models.Discount'
        type: array
      end_date:
        type: string
      id:
        type: string
      is_cancelled:
        type: boolean
      is_featured:
        description: Featured event flag
        type: boolean
      location:
        description: Keep for backward compatibility
        type: string
      organizer:
        $ref: '#/definitions/models.User'
      organizer_id:
        type: string
      price:
        description: Base price for backward compatibility
        minimum: 0
        type: number
      promocodes:
        items:
          $ref: '#/definitions/models.Promocode'
        type: array
      sales_status:
        description: active, paused, stopped
        type: string
      start_date:
        type: string
      status:
        description: draft, pending, approved, held, rejected, cancelled
        type: string
      tiers:
        description: Relations
        items:
          $ref: '#/definitions/models.EventTier'
        type: array
      timezone:
        type: string
      title:
        type: string
      updated_at:
        type: string
      venue_name:
        type: string
    required:
    - capacity
    - end_date
    - price
    - start_date
    - title
    type: object
...
  models.Promocode:
    properties:
      amount:
        description: Percentage or fixed amount
        type: number
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_discount:
        description: Maximum discount amount (0 = unlimited)
        type: number
      min_amount:
        description: Minimum order amount
        type: number
      name:
        type: string
      tier:
        $ref: '#/definitions/models.EventTier'
      tier_id:
        description: If promocode applies to specific tier
        type: string
      type:
        description: percentage, fixed_amount
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
      usage_limit:
        description: 0 = unlimited
        type: integer
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
...
  models.Ticket:
    properties:
      check_in_time:
        type: string
      check_out_time:
        type: string
      checked_in_by:
        type: string
      checked_out_by:
        type: string
      created_at:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        type: string
      purchase_date:
        type: string
      quantity:
        type: integer
      status:
        description: active, used, cancelled, refunded
        type: string
      ticket_number:
        description: Unique ticket identifier
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TicketCheckInRequest:
    properties:
      event_id:
        type: string
      ticket_number:
        type: string
    required:
    - event_id
    - ticket_number
    type: object
  models.TicketCheckOutRequest:
    properties:
      event_id:
        type: string
      ticket_number:
        type: string
    required:
    - event_id
    - ticket_number
    type: object
  models.TicketResponse:
    properties:
      check_in_time:
        type: string
      check_out_time:
        type: string
      checked_in_by:
        type: string
      checked_out_by:
        type: string
      created_at:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        type: string
      purchase_date:
        type: string
      quantity:
        type: integer
      status:
        type: string
      ticket_number:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
      user_id:
        type: string
    type: object
  models.TicketScanResponse:
    properties:
      check_in_time:
        type: string
      check_out_time:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      scan_time:
        type: string
      scanned_by:
        $ref: '#/definitions/models.UserResponse'
      status:
        type: string
      ticket_number:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
...
  models.User:
    properties:
      account_status:
        description: active, inactive, suspended
        type: string
      admin_remark:
        type: string
      country_code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_email_verified:
        type: boolean
      last_name:
        type: string
      organization:
        $ref: '#/definitions/models.Organization'
      organization_id:
        type: string
      organizer_status:
        description: inactive, pending, approved, rejected
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      account_status:
        type: string
      admin_remark:
        type: string
      country_code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_email_verified:
        type: boolean
      last_name:
        type: string
      organization:
        $ref: '#/definitions/models.OrganizationResponse'
      organization_id:
        type: string
      organizer_status:
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.RoleResponse'
        type: array
      updated_at:
        type: string
    type: object
  utils.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      fields:
        description: For validation errors
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/utils.ErrorInfo'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A scalable REST API for event ticketing system with secure authentication
    and role-based access control (RBAC)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Timro Ticket API
  version: "1.0"
paths:
...
  /api/v1/organizer/events/{id}/tickets:
    get:
      description: Get all tickets purchased for a specific event (Organizer API)
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.TicketResponse'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get event tickets
      tags:
      - Organizer
  /api/v1/organizer/events/{id}/tickets/stats:
    get:
      description: Get ticket statistics for a specific event (Organizer API)
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get ticket statistics
      tags:
      - Organizer
...
  /api/v1/organizer/tickets/checkin:
    post:
      consumes:
      - application/json
      description: Mark a ticket as checked-in for an event (Organizer API)
      parameters:
      - description: Check-in details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TicketCheckInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.TicketResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Check-in ticket
      tags:
      - Organizer
  /api/v1/organizer/tickets/checkout:
    post:
      consumes:
      - application/json
      description: Mark a ticket as checked-out from an event (Organizer API)
      parameters:
      - description: Check-out details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TicketCheckOutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.TicketResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Check-out ticket
      tags:
      - Organizer
  /api/v1/organizer/tickets/scan:
    post:
      consumes:
      - application/json
      description: Scan a ticket QR code to check-in or check-out attendee (Organizer
        API)
      parameters:
      - description: Ticket scan details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TicketCheckInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.TicketScanResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Scan ticket for check-in/check-out
      tags:
      - Organizer
...
  /api/v1/user/events/{event_id}/tickets:
    get:
      description: Get all tickets purchased by the user for a specific event
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user's tickets for a specific event
      tags:
      - User Tickets
  /api/v1/user/tickets:
    get:
      description: Get list of tickets purchased by the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by ticket status (active, used, cancelled)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user's purchased tickets
      tags:
      - User Tickets
  /api/v1/user/tickets/{id}:
    get:
      description: Get detailed information about a specific ticket owned by the user
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Ticket'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get specific ticket details
      tags:
      - User Tickets
  /api/v1/user/tickets/{id}/qr:
    get:
      description: Get QR code data for a specific ticket (for mobile apps)
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get ticket QR code
      tags:
      - User Tickets
  /api/v1/user/tickets/stats:
    get:
      description: Get statistics about user's ticket purchases and usage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user's ticket statistics
      tags:
      - User Tickets
securityDefinitions:
  ApiKeyAuth:
    description: 'JWT token authentication. Use the ''Bearer'' prefix followed by
      a space and the access token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"